# BitWaves.WebAPI/docs/swagger/bitwaves-v1.yaml
swagger: '2.0'
info:
  description: BitWaves 后端接口文档
  version: 1.0.0
  title: BitWaves Web API Docs
  contact:
    email: msrlancern@126.com
...
paths:
  /contents:
    get:
      summary: 获取静态对象列表
      description: 获取静态对象列表
      tags:
        - contents
      parameters:
        - in: header
          name: Authorization
          required: true
          type: string
          description: 身份验证标识，格式为 `Jwt <jwt>`。注意 `Jwt` 与 `<jwt>` 之间的空格
        - in: query
          name: name
          required: false
          type: string
          description: 要筛选的静态对象的名称
        - in: query
          name: mimeType
          required: false
          type: string
          description: 要筛选的静态对象的 MIME 类型
        - in: query
          name: page
          required: false
          type: integer
          description: 分页参数，页面编号。页面编号从 0 开始，默认为 0
        - in: query
          name: itemsPerPage
          required: false
          type: integer
          description: 分页参数，每个页面中的元素数量
      responses:
        200:
          description: 操作成功完成
          headers:
            X-BitWaves-Count:
              type: integer
              description: 分页前满足筛选条件的静态对象数量
          schema:
            type: array
            items:
              $ref: '#/definitions/ContentObjectInfo'
        400:
          description: 无效的查询参数
        403:
          description: 用户没有相应的权限
    post:
      summary: 上传一个静态对象
      description: 上传一个静态对象。静态对象的大小不应该超过 14 MB
      tags:
        - contents
      parameters:
        - in: header
          name: Authorization
          required: true
          type: string
          description: 身份验证标识，格式为 `Jwt <jwt>`。注意 `Jwt` 与 `<jwt>` 之间的空格
        - in: formData
          required: true
          name: content
          type: file
          description: 包含静态对象数据的文件。文件大小不应该超过 14 MB
      responses:
        200:
          description: 操作成功完成
          schema:
            type: object
            required:
              - id
            properties:
              id:
                type: string
                description: 新创建的静态对象的全局唯一 ID
                example: a154edb2431fdbb230fbf384
        400:
          description: 无效的输入数据
        403:
          description: 用户没有相应的权限
  /contents/{contentId}:
    get:
      summary: 获取静态对象数据
      description: 获取静态对象数据
      tags:
        - contents
      parameters:
        - in: path
          name: contentId
          type: string
          required: true
          description: 静态对象的全局唯一 ID
        - in: query
          name: attachment
          type: boolean
          required: false
          description: 若该参数被设置为 `true`，则返回的响应中包含 `Content-Disposition` 头部字段且被设置为 `attachment`。通常情况下这样的响应会直接拉起浏览器的"下载"对话框询问用户保存地址。默认为 `false`
      responses:
        200:
          description: 操作成功完成
        400:
          description: 无效的请求参数
        404:
          description: 无法找到指定的静态对象
    delete:
      summary: 删除静态对象
      description: 删除静态对象
      tags:
        - contents
      parameters:
        - in: path
          name: contentId
          required: true
          type: string
          description: 要删除的静态对象的全局唯一 ID
        - in: header
          name: Authorization
          required: true
          type: string
          description: 身份验证标识，格式为 `Jwt <jwt>`。注意 `Jwt` 与 `<jwt>` 之间的空格
      responses:
        200:
          description: 操作成功完成
        400:
          description: 无效的静态对象 ID
        403:
          description: 用户没有相应的权限
        404:
          description: 未能找到指定的静态对象
...
  /announcements:
    get:
      summary: 获取全站公告列表
      description: 获取全站公告列表
      tags:
        - announcements
      parameters:
        - in: query
          name: page
          type: integer
          description: 分页参数，页面编号。页面编号从 0 开始，默认为 0
        - in: query
          name: itemsPerPage
          type: integer
          description: 分页参数，每页上的元素数量，默认为 20
      responses:
        200:
          description: 操作成功完成
          headers:
            X-BitWaves-Count:
              type: integer
              description: 分页前满足筛选条件的公告数量
          schema:
            type: array
            items:
              $ref: '#/definitions/AnnouncementListItem'
        400:
          description: 无效的查询参数
    post:
      description: 创建新的全站公告
      summary: 创建新的全站公告
      tags:
        - announcements
      parameters:
        - in: header
          name: Authorization
          type: string
          required: true
          description: 身份验证标识，格式为 `Jwt <jwt>`。注意 `Jwt` 与 `<jwt>` 之间的空格
        - in: body
          name: data
          required: true
          description: 要创建的公告信息
          schema:
            $ref: '#/definitions/PostAnnouncementModel'
      responses:
        200:
          description: 操作成功完成
          schema:
            type: object
            required:
              - id
            properties:
              id:
                type: string
                description: 新创建的全站公告的全局唯一 ID
                example: a154edb2431fdbb230fbf384
        400:
          description: 无效的请求参数
        403:
          description: 用户没有相应的权限
  /announcements/{announcementId}:
    get:
      summary: 获取全站公告信息
      description: 获取全站公告信息
      tags:
        - announcements
      parameters:
        - in: path
          name: announcementId
          type: string
          required: true
          description: 公告的 ID
      responses:
        200:
          description: 操作成功完成
          schema:
            $ref: '#/definitions/AnnouncementInfo'
        400:
          description: 无效的公告 ID
        404:
          description: 无法找到给定的公告
    put:
      summary: 修改一个全站公告
      description: 修改一个全站公告
      tags:
        - announcements
      parameters:
        - in: path
          name: announcementId
          type: string
          required: true
          description: 要修改的全站公告 ID
        - in: header
          name: Authorization
          type: string
          required: true
          description: 身份验证标识，格式为 `Jwt <jwt>`。注意 `Jwt` 与 `<jwt>` 之间的空格
        - in: body
          name: data
          required: true
          schema:
            $ref: '#/definitions/PutAnnouncementModel'
      responses:
        200:
          description: 操作成功完成
        400:
          description: 无效的请求参数
        403:
          description: 用户没有相应的权限
        404:
          description: 未能找到指定的全站公告
    delete:
      summary: 删除一个全站公告
      description: 删除一个全站公告
      tags:
        - announcements
      parameters:
        - in: path
          name: announcementId
          type: string
          required: true
          description: 要修改的全站公告 ID
        - in: header
          name: Authorization
          type: string
          required: true
          description: 身份验证标识，格式为 `Jwt <jwt>`。注意 `Jwt` 与 `<jwt>` 之间的空格
      responses:
        200:
          description: 操作成功完成
        400:
          description: 无效的请求参数
        403:
          description: 用户没有相应的权限
        404:
          description: 无法找到指定的全站公告
...
  /archive:
    get:
      summary: 获取 Problem Archive 列表
      description: 获取 Problem Archive 列表
      tags:
        - archive
      produces:
        - application/json
      parameters:
        - in: query
          name: by
          type: string
          required: false
          description: 列表排序方法
          enum:
            - Id
            - Difficulty
            - LastUpdateTime
            - TotalSubmissions
            - AcceptedSubmissions
            - TotalAttemptedUsers
            - TotalSolvedUsers
          default: Id
        - in: query
          name: descend
          type: boolean
          required: false
          description: 是否按照给定关键字的降序排列
          default: false
        - in: query
          name: page
          type: integer
          required: false
          description: 分页参数，要返回的列表的页面编号
          minimum: 0
          default: 0
        - in: query
          name: itemsPerPage
          type: integer
          required: false
          description: 分页参数，每一页上的项目数量
          minimum: 1
          default: 20
      responses:
        200:
          description: 操作成功完成
          headers:
            X-BitWaves-Count:
              type: integer
              description: 分页前满足筛选条件的题目数量
          schema:
            type: array
            items:
              $ref: '#/definitions/ProblemListItem'
        400:
          description: 请求中携带的参数无效
    post:
      summary: 向公开题目集中添加题目
      description: 向公开题目集中添加题目
      tags:
        - archive
      parameters:
        - in: header
          name: Authorization
          type: string
          required: true
          description: 用户身份验证标识，内容格式为 `Jwt <jwt>`。注意 `Jwt` 与 `<jwt>` 之间的空格
        - in: body
          name: data
          required: true
          description: 要添加的题目数据
          schema:
            $ref: '#/definitions/PostArchiveModel'
      responses:
        200:
          description: 操作成功完成
        400:
          description: 请求中的数据不合法
        403:
          description: 用户没有权限执行此操作
        404:
          description: 无法找到给定的题目
        409:
          description: 公开题目集 ID 冲突
    delete:
      summary: 从公开题目集中删除题目
      description: 从公开题目集中删除题目
      tags:
        - archive
      parameters:
        - in: header
          name: Authorization
          type: string
          required: true
          description: 用户身份验证标识，内容格式为 `Jwt <jwt>`。注意 `Jwt` 与 `<jwt>` 之间的空格
        - in: body
          name: data
          required: true
          description: 要删除的题目在公开题目集中的 ID
          schema:
            type: array
            items:
              type: integer
            example:
              - 1
              - 3
              - 200
      responses:
        200:
          description: 操作成功完成
        400:
          description: 请求中的数据无效
        403:
          description: 用户没有权限执行此操作
  /archive/{archiveId}:
    get:
      summary: 获取公开题目集中给定 ID 题目的详细信息
      description: 获取公开题目集中给定 ID 题目的详细信息
      tags:
        - archive
      parameters:
        - in: path
          name: archiveId
          required: true
          type: integer
          description: 题目在公开题目集中的 ID
      responses:
        200:
          description: 操作成功完成
          schema:
            $ref: '#/definitions/ProblemInfo'
        400:
          description: 无效的题目 ID
        404:
          description: 无法找到指定的题目
...
definitions:
  ContentObjectInfo:
    type: object
    required:
      - id
      - name
      - mimeType
      - creationTime
      - size
    properties:
      id:
        type: string
        description: 静态对象的全局唯一 ID
        example: a154edb2431fdbb230fbf384
      name:
        type: string
        description: 静态对象的名称
        example: bit-logo.jpg
      mimeType:
        type: string
        description: 静态对象的 MIME 类型
        example: image/jpg
      creationTime:
        type: string
        description: 静态对象被创建时的时间，UTC 时区
        example: 2019-10-04T14:23:56.241Z
      size:
        type: integer
        description: 静态对象的字节大小
        example: 572243
...
  PostAnnouncementModel:
    type: object
    required:
      - title
      - content
    properties:
      title:
        type: string
        minLength: 1
        maxLength: 128
        description: 全站公告的标题
        example: BITOJ 即日起开始营业
      pinned:
        type: boolean
        description: 公告是否置顶，默认为 false
        example: false
      content:
        type: string
        minLength: 1
        description: 全站公告的内容
        example: BITOJ 即日起开始营业，欢迎 AK！
  PutAnnouncementModel:
    type: object
    properties:
      title:
        type: string
        minLength: 1
        maxLength: 128
        description: 修改后的标题。若不需要修改标题，请不要提供此成员
      pinned:
        type: boolean
        description: 修改后的置顶标记。若不需要修改此属性，请不要提供此成员
      content:
        type: string
        minLength: 1
        description: 修改后的内容。若不需要修改此属性，请不要提供此成员
  AnnouncementListItem:
    type: object
    required:
      - id
      - title
      - author
      - creationTime
      - lastUpdateTime
      - pinned
    properties:
      id:
        type: string
        description: 全站公告的 ID
        example: 5d8825e87d2847a7e4886fbf
      title:
        type: string
        description: 全站公告的标题
        example: BITOJ 即日起开始营业
      author:
        type: string
        description: 全站公告的作者的用户名
        example: admin
      creationTime:
        type: string
        description: 全站公告的创建时间
        example: "2019-11-01T13:25:37.371Z"
      lastUpdateTime:
        type: string
        description: 全站公告的上次修改时间
        example: "2019-11-01T13:25:37.371Z"
      pinned:
        type: boolean
        description: 公告是否置顶
        example: false
  AnnouncementInfo:
    type: object
    required:
      - id
      - title
      - author
      - creationTime
      - lastUpdateTime
      - pinned
      - content
    properties:
      id:
        type: string
        description: 全站公告的 ID
        example: 5d8825e87d2847a7e4886fbf
      title:
        type: string
        description: 全站公告的标题
        example: BITOJ 即日起开始营业
      author:
        type: string
        description: 全站公告的作者的用户名
        example: admin
      creationTime:
        type: string
        description: 全站公告的创建时间
        example: "2019-11-01T13:25:37.371Z"
      lastUpdateTime:
        type: string
        description: 全站公告的上次修改时间
        example: "2019-11-01T13:25:37.371Z"
      pinned:
        type: boolean
        description: 公告是否置顶
        example: false
      content:
        type: string
        description: 全站公告的内容
        example: BITOJ 即日起开始营业，欢迎 AK！
...
  ProblemListItem:
    type: object
    required:
      - id
      - archiveId
      - title
      - creationTime
      - lastUpdateTime
      - author
      - source
      - difficulty
      - tags
      - totalSubmissions
      - acceptedSubmissions
      - totalAttemptedUsers
      - totalSolvedUsers
      - lastSubmissionTime
    properties:
      id:
        type: string
        description: 题目的全局唯一 ID，请注意与题目的 archiveId 的区别
        example: "5d8825e87d2847a7e4886fbf"
      archiveId:
        type: integer
        description: 题目在 Problem Archive 中的 ID
        example: 10
      title:
        type: string
        description: 题目的标题
        example: "Two Sums"
      creationTime:
        type: string
        description: 题目的创建时间，UTC 时区
        example: "2019-09-23T01:54:48.087Z"
      lastUpdateTime:
        type: string
        description: 题目上次更新的时间，UTC 时区
        example: "2019-09-23T01:54:48.087Z"
      author:
        type: string
        description: 题目的作者的用户名
        example: "admin"
      source:
        type: string
        description: 题目的来源
        example: "2018年北京理工大学程序设计校赛"
      difficulty:
        type: integer
        description: 题目的难度系数
        example: 80
      tags:
        type: array
        items:
          type: string
        description: 题目的标签
        example:
          - 字符串
          - 模拟
          - 数据结构
      totalSubmissions:
        type: integer
        description: 题目的总提交数量
        example: 1000
      acceptedSubmissions:
        type: integer
        description: 题目的 AC 提交数量
        example: 100
      totalAttemptedUsers:
        type: integer
        description: 尝试过该题目的用户总数
        example: 100
      totalSolvedUsers:
        type: integer
        description: 成功解答该题目的用户总数
        example: 40
      lastSubmissionTime:
        type: string
        description: 题目的上次提交时间
        example: "2019-09-23T01:54:48.087Z"
  ProblemInfo:
    type: object
    required:
      - id
      - archiveId
      - title
      - creationTime
      - lastUpdateTime
      - author
      - source
      - legend
      - input
      - output
      - samples
      - notes
      - difficulty
      - tags
      - timeLimit
      - memoryLimit
      - judgeMode
      - totalSubmissions
      - acceptedSubmissions
      - testReady
    properties:
      id:
        type: string
        description: 题目的全局唯一 ID
        example: "5d8825e87d2847a7e4886fbf"
      archiveId:
        type: integer
        description: 题目在公开题目集中的 ID。若题目不在公开题目集中，该字段为 null
      title:
        type: string
        description: 题目标题
        example: "诸神眷顾的幻想乡"
      creationTime:
        type: string
        description: 题目的创建时间，UTC 时区
        example: "2019-09-23T01:54:48.087Z"
      lastUpdateTime:
        type: string
        description: 题目上次更新的时间，UTC 时区
        example: "2019-09-23T01:54:48.087Z"
      author:
        type: string
        description: 题目的作者的用户名
        example: "admin"
      source:
        type: string
        description: 题目的来源
        example: "2018年北京理工大学程序设计校赛"
      legend:
        type: string
        description: 题目正文叙述
        example: "hello world"
      input:
        type: string
        description: 题目输入描述
        example: "input description"
      output:
        type: string
        description: 题目输出描述
        example: "output description"
      samples:
        type: array
        description: 题目样例
        items:
          type: object
          properties:
            input:
              type: string
              description: 样例输入
              example: "1 2"
            output:
              type: string
              description: 样例输出
              example: "3"
      notes:
        type: string
        description: 题目提示信息
        example: "notes"
      difficulty:
        type: integer
        description: 题目难度系数，范围为 [0, 100]
        example: 80
      tags:
        type: array
        description: 题目标签
        items:
          type: string
        example:
          - "字符串"
          - "模拟"
          - "数据结构"
      timeLimit:
        type: integer
        description: 题目时间限制，单位为毫秒
        example: 1000
      memoryLimit:
        type: integer
        description: 题目内存限制，单位为 MB
        example: 256
      judgeMode:
        type: string
        description: 题目评测模式
        enum:
          - Standard
          - SpecialJudge
          - Interactive
      builtinCheckerOptions:
        type: array
        items:
          type: string
        description: 当评测模式为 Standard 时，该字段表示内建答案检查器的选项。可能的值有 `FloatingPointAware` 以及 `IgnoreCase`
        example:
          - FloatingPointAware
          - IgnoreCase
      totalSubmissions:
        type: integer
        description: 题目总提交数量
        example: 100
      acceptedSubmissions:
        type: integer
        description: 题目 AC 提交数量
        example: 80
      totalAttemptedUsers:
        type: integer
        description: 尝试过该题目的用户总数
        example: 100
      totalSolvedUsers:
        type: integer
        description: 成功解答该题目的用户总数
        example: 40
      lastSubmissionTime:
        type: string
        description: 题目的上次提交时间
        example: "2019-09-23T01:54:48.087Z"
      testReady:
        type: boolean
        description: 用户是否可以提交答案
        example: true
  PostArchiveModel:
    type: object
    required:
      - id
      - archiveId
    properties:
      id:
        type: string
        description: 题目的全局唯一 ID
        example: "5d8825e87d2847a7e4886fbf"
      archiveId:
        type: integer
        description: 题目在公开题目集中的 ID
        minimum: 0
        example: 100
...